Moved RegionManagerAwareBehavior to Prism.Regions.Behaviors namespace and Register RegionManagerAwareBehavior in both Bootstrapper and PrismApplicationBase
